pyproject.toml
[tool.poetry]
name = "main"
version = "0.1.0"
description = "Example app"
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110"
uvicorn = "^0.29"
celery = "^5.5.2"
redis = "^6.0.0"
python-dotenv = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
mypy = "^1.9"
pre-commit = "^3.7"
httpx = "^0.27"



